requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
tqdm==4.66.4
scikit-learn==1.5.1
numpy==1.26.4
joblib==1.4.2
